+++ src/features/abort-error/abort-error.types.ts
@@ -6,6 +6,5 @@ export interface AbortErrorOptions {
   type?: AbortType;
   reason?: any;
   cause?: Error;
-  signal?: AbortSignal;
   initiator?: AbortInitiator;
 }